Safari-609.1.11.1/LayoutTests/js/dom/dfg-patchable-get-by-id-after-watchpoint-expected.txt

This tests that a patchable GetById right after a watchpoint has the appropriate nop padding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS foo(o, p) is 27
PASS foo(o, p) is 44
PASS successfullyParsed is true

TEST COMPLETE
